// src/ensembl/src/shared/dropdown/dropdown-types.ts
export enum Position {
  TOP_LEFT = 'top_left',
  TOP_CENTRE = 'top_centre',
  TOP_RIGHT = 'top_right',
  RIGHT_TOP = 'right_top',
  RIGHT_CENTRE = 'right_centre',
  RIGHT_BOTTOM = 'right_bottom',
  BOTTOM_RIGHT = 'bottom_right',
  BOTTOM_CENTRE = 'bottom_centre',
  BOTTOM_LEFT = 'bottom_left',
  LEFT_BOTTOM = 'left_bottom',
  LEFT_CENTRE = 'left_centre',
  LEFT_TOP = 'left_top'
}